requirements.txt
pandas==1.5.3
openpyxl==3.1.2
fpdf2==2.4.5
